requirements.txt
pymongo==4.10.1
requests==2.32.3
fastapi[standard]==0.115.5
PyJWT==2.8.0
cryptography==42.0.5